replaced old server system with new struct-based version
