+++ bb84/remote_timetagger.py
@@ -1,8 +1,12 @@
 import sys
 import os
 import socket
+import struct
+import time
 import typing
-import json
+import pprint
+
+import numpy
 
 sys.path.append(
     os.path.abspath(os.path.join(
@@ -11,101 +15,98 @@ sys.path.append(
     ))
 )
 import bb84.timetagger as timetagger
+import bb84.remote_timetagger_protocol as remote_timetagger_protocol
 
 server_host = '127.0.0.1'
-# server_host = '137.195.89.222'
+server_host = '137.195.63.6'
 server_port = 5003
 
-def send_request(
-        host: str,
-        port: int,
-        command: str,
-        arguments: list = [],
-        timeout: int = 5
-) -> typing.Any:
-    request = {'command': command}
-    try:
-        with socket.create_connection(
-            address=(host, port),
-            timeout=timeout
-        ) as s:
-            match command:
-                case 'list_devices':
-                    pass
-
-                case 'measure':
-                    pass
-
-                case _:
-                    raise Exception('Unknown command')
-
-            s.sendall(json.dumps(request).encode())
-            buffer = ""
-            while True:
-                data = s.recv(1024).decode()
-                if not data:
-                    break
-                buffer += data
-                if "\n" in buffer:
-                    break
-            response = json.loads(buffer.strip())
-            return response
-
-    except Exception as e:
-        return {f'Error sending request {request}': str(e)}
-    
-def list_device_info(
-        host: str,
-        port: int
-) -> list[timetagger.DeviceInfo]:
-    result = send_request(
-        host=host,
-        port=port,
-        command='list_devices'
-    )
-    return [timetagger.DeviceInfo(**dev) for dev in result['devices']]
+def parse_status(payload: bytes) -> str:
+    message_len = struct.unpack('I', payload[:4])[0]
+    message = struct.unpack(f'{message_len}s', payload[4:])[0]
+    return bytes(message).decode()
 
 class Timetagger(timetagger.TimeTagger):
     def __init__(
             self,
             host: str,
-            port: int,
+            port: int
     ) -> None:
         self.host = host
         self.port = port
 
-        self._get_timetagger()
-
-    def _get_timetagger(
-            self
-    ) -> None:
-        devices = list_device_info(
-            host=self.host,
-            port=self.port
+        self._sock = socket.socket(
+            socket.AF_INET,
+            socket.SOCK_STREAM
         )
-        # dev_index = next(
-        #     (i for i, dev in enumerate(devices) if str(dev.serial_number) == serial_number),
-        #     None
-        # )
-        # if dev_index is None:
-        #     raise Exception
-        # else:
-        #     self.device_info = devices[dev_index]
+        self._sock.connect((self.host, self.port))
+        self._get_device_info()
 
-        self.device_info = devices[0]
+    def __del__(self) -> None:
+        self.disconnect()
 
     def measure(self) -> timetagger.RawData:
-        result = send_request(
-            host=self.host,
-            port=self.port,
-            command='measure'
+        self._send_command(
+            command=remote_timetagger_protocol.Command.MEASURE_ONCE
         )
-        rawdata = timetagger.RawData(**result['rawdata'])
-        return rawdata
+        resp_type, payload = self._receive_response()
+        if resp_type == remote_timetagger_protocol.Response.RAWDATA:
+            return timetagger.RawData.deserialise(
+                payload=payload
+            )
+        else:
+            print('Unexpected response:', resp_type)
+            return timetagger.RawData(
+                timetags=numpy.empty(0),
+                channels=numpy.empty(0)
+            )
+
+    def disconnect(self) -> None:
+        self._sock.close()
+
+    def _get_device_info(self) -> None:
+        self._send_command(
+            command=remote_timetagger_protocol.Command.LIST_DEVICES
+        )
+        resp_type, payload = self._receive_response()
+        if resp_type == remote_timetagger_protocol.Response.DEVICE_INFO:
+            self.device_info=timetagger.DeviceInfo.deserialise(
+                payload=payload
+            )
+        else:
+            print('Unexpected response:', resp_type)
+
+    def _send_command(
+            self,
+            command: remote_timetagger_protocol.Command
+        ) -> None:
+        self._sock.sendall(struct.pack('I', command))
+
+    def _recvall(self, size: int) -> bytes:
+        data = bytearray()
+        while len(data) < size:
+            part = self._sock.recv(size - len(data))
+            if not part:
+                raise ConnectionError('Socket closed')
+            data.extend(part)
+        return data
+
+    def _receive_response(self) -> tuple[typing.Any, bytes]:
+        header = self._recvall(size=5)
+        total_len, resp_type = struct.unpack('IB', header)
+        payload = self._recvall(total_len - 1)
+        return resp_type, payload
 
 if __name__ == '__main__':
     tt = Timetagger(
         host=server_host,
         port=server_port
     )
-    print(timetagger.Data().from_raw_data(tt.measure()))
+    print(tt.device_info)
+    # for _ in range(10):
+    #     raw_data = tt.measure()
+    #     # print(timetagger.Data.from_raw_data(raw_data=raw_data))
+    #     # singles = numpy.bincount(raw_data.channels, minlength=8)
+    #     # print(singles)
+    #     pprint.pprint(raw_data)
+    #     time.sleep(1)
