+++ polarimeter/remote_polarimeter.py
@@ -1,8 +1,7 @@
 import sys
 import os
 import socket
-import typing
-import json
+import struct
 
 sys.path.append(
     os.path.abspath(os.path.join(
@@ -11,59 +10,16 @@ sys.path.append(
     ))
 )
 import polarimeter.thorlabs_polarimeter as thorlabs_polarimeter
+import polarimeter.remote_polarimeter_protocol as remote_polarimeter_protocol
 
 server_host = '127.0.0.1'
-# server_host = '137.195.89.222'
+server_host = '137.195.89.222'
 server_port = 5003
 
-def send_request(
-        host: str,
-        port: int,
-        command: str,
-        arguments: list = [],
-        timeout: int = 5
-) -> typing.Any:
-    request = {'command': command}
-    try:
-        with socket.create_connection(
-            address=(host, port),
-            timeout=timeout
-        ) as s:
-            match command:
-                case 'list_devices':
-                    pass
-
-                case 'measure':
-                    pass
-
-                case _:
-                    raise Exception('Unknown command')
-
-            s.sendall(json.dumps(request).encode())
-            buffer = ""
-            while True:
-                data = s.recv(1024).decode()
-                if not data:
-                    break
-                buffer += data
-                if "\n" in buffer:
-                    break
-            response = json.loads(buffer.strip())
-            return response
-
-    except Exception as e:
-        return {f'Error sending request {request}': str(e)}
-    
-def list_device_info(
-        host: str,
-        port: int
-) -> list[thorlabs_polarimeter.DeviceInfo]:
-    result = send_request(
-        host=host,
-        port=port,
-        command='list_devices'
-    )
-    return [thorlabs_polarimeter.DeviceInfo(**dev) for dev in result['devices']]
+def parse_status(payload: bytes) -> str:
+    message_len = struct.unpack('I', payload[:4])[0]
+    message = struct.unpack(f'{message_len}s', payload[4:])[0]
+    return bytes(message).decode()
 
 class Polarimeter(thorlabs_polarimeter.Polarimeter):
     def __init__(
@@ -75,33 +31,66 @@ class Polarimeter(thorlabs_polarimeter.Polarimeter):
         self.host = host
         self.port = port
 
-        self._get_polarimeter(serial_number=serial_number)
+        self._sock = socket.socket(
+            socket.AF_INET,
+            socket.SOCK_STREAM
+        )
+        self._sock.connect((self.host, self.port))
+        self._get_device_info(serial_number=serial_number)
 
-    def _get_polarimeter(
+    def __del__(self) -> None:
+        self.disconnect()
+
+    def measure(self) -> thorlabs_polarimeter.RawData:
+        self._send_command(
+            command=remote_polarimeter_protocol.Command.MEASURE_ONCE
+        )
+        resp_type, payload = self._receive_response()
+        if resp_type == remote_polarimeter_protocol.Response.RAWDATA:
+            return thorlabs_polarimeter.RawData.deserialise(
+                payload=payload
+            )
+        else:
+            print('Unexpected response:', resp_type)
+
+    def disconnect(self) -> None:
+        self._sock.close()
+
+    def _get_device_info(
             self,
             serial_number: str
     ) -> None:
-        devices = list_device_info(
-            host=self.host,
-            port=self.port
+        self._send_command(
+            command=remote_polarimeter_protocol.Command.LIST_DEVICES
         )
-        dev_index = next(
-            (i for i, dev in enumerate(devices) if str(dev.serial_number) == serial_number),
-            None
-        )
-        if dev_index is None:
-            raise Exception
+        resp_type, payload = self._receive_response()
+        if resp_type == remote_polarimeter_protocol.Response.DEVICE_INFO:
+            self.device_info=thorlabs_polarimeter.DeviceInfo.deserialise(
+                payload=payload
+            )
         else:
-            self.device_info = devices[dev_index]
+            print('Unexpected response:', resp_type)
 
-    def measure(self) -> thorlabs_polarimeter.RawData:
-        result = send_request(
-            host=self.host,
-            port=self.port,
-            command='measure'
-        )
-        rawdata = thorlabs_polarimeter.RawData(**result['rawdata'])
-        return rawdata
+    def _send_command(
+            self,
+            command: remote_polarimeter_protocol.Command
+        ) -> None:
+        self._sock.sendall(struct.pack('I', command))
+
+    def _recvall(self, size: int) -> bytes:
+        data = bytearray()
+        while len(data) < size:
+            part = self._sock.recv(size - len(data))
+            if not part:
+                raise ConnectionError('Socket closed')
+            data.extend(part)
+        return data
+
+    def _receive_response(self) -> tuple[int, bytes]:
+        header = self._recvall(size=5)
+        total_len, resp_type = struct.unpack('IB', header)
+        payload = self._recvall(total_len - 1)
+        return resp_type, payload
 
 if __name__ == '__main__':
     pax = Polarimeter(
@@ -109,4 +98,4 @@ if __name__ == '__main__':
         port=server_port,
         serial_number='M00910360'
     )
-    print(thorlabs_polarimeter.Data().from_raw_data(raw_data=pax.measure()))
+    print(pax.device_info)
